requirements-base.txt
SQLAlchemy
psycopg2-binary
alembic
python-dotenv
fastapi
websockets
coloredlogs
confluent-kafka